[CQ-W9-CLIPPY-CHURN] fix rust-1.91 clippy errors so -D warnings is green (audit-rust §7)
Rust 1.91 added/promoted several clippy lints that broke
`cargo clippy --workspace --all-features --all-targets -- -D warnings`.
This commit applies pure idiom rewrites; no behavior changes.

Errors fixed:
- updater/src/main.rs:1080,1088: suspicious_open_options
  Add explicit .truncate(false) to lock-file OpenOptions chains
  (lock semantics preserved - we never wanted to truncate).
- indi/src/{dome,filterwheel,focuser,mount,rotator}.rs:
  field_reassign_with_default - rewrite as struct-update syntax
  IndiTimeoutConfig { field: x, ..Default::default() }.
- indi/src/client.rs:3057: manual_range_contains
  Rewrite as (9900..=11000).contains(&time_since).
- imaging/src/stats.rs:1105,1151: field_reassign_with_default
  Rewrite StarDetectionConfig as struct-update syntax.
- imaging/src/platesolve.rs:1675: ptr_arg
  Change test helper signature &PathBuf -> &Path.
- imaging/src/stacking.rs:693: needless_range_loop
  Replace index-loop with .iter_mut().enumerate().
- sequencer/src/all_sky_polar.rs:672: too_many_arguments
  Test-only physics simulator helper - add #[allow] with rationale.
- alpaca/src/camera.rs:1986: too_many_arguments
  Test-only ImageBytes payload builder - add #[allow] with rationale.
- sequencer/src/meridian_flip_executor.rs:1822,1855,1977,2019:
  field_reassign_with_default - rewrite MeridianFlipConfig as
  struct-update syntax.

Auto-fixed by `cargo clippy --fix` (idiomatic rewrites only):
- imaging/src/debayer.rs, processing.rs, fits.rs (identity ops,
  repeat -> repeat_n, range_contains in test assertions).
- sequencer/src/{autofocus,focus_prediction,meridian}.rs
  (range_contains in test assertions, vec![] -> [], identity casts).
- alpaca/src/guard.rs (unused use super::*; in stub test module).

Bridge-module rustc lints surfaced once compilation got past indi:
- bridge/src/api/mod.rs: change pub mod {connection,heartbeat,
  api_version,storage} to pub(crate) mod so the pub use api::*
  glob in lib.rs no longer collides with pub use device_manager::*
  and the local mod storage. Internal use via crate::api::xxx and
  frb_generated.rs::crate::api::xxx paths still works (those use
  crate-relative paths). Fixes hidden_glob_reexports +
  ambiguous_glob_reexports errors.

Verify:
  cd native/nightshade_native
  cargo clippy --workspace --all-features --all-targets -- \
    -D warnings -D clippy::result_unit_err -D clippy::await_holding_lock
  -> Finished `dev` profile, exit 0
  cargo test --workspace --all-features --no-run -> compiles clean.

+++ native/nightshade_native/imaging/src/processing.rs
@@ -579,7 +579,7 @@ mod tests {
         let mut data = Vec::with_capacity((w * h * 2) as usize);
         for _y in 0..h {
             for x in 0..w {
-                let v = ((x as u32 * 65535) / (w - 1)) as u16;
+                let v = ((x * 65535) / (w - 1)) as u16;
                 data.extend_from_slice(&v.to_le_bytes());
             }
         }
